Multi-threaded: using threads here "samtools view" run as a single-cmd
NOTE: when it's in a compound command, e.g. in a pipe or &&, we leave it
single-threaded b/c the consumption of more threads for the rule doesn't seem
to benefit.  NEED to test those situations out more.

NOTE: with these simple improvements, we were able to gain 8.77% efficiency in
processing time (on test_runs/GATA3_chilin_comp see 2017-07-24_todo.txt) over
base-line!